+++ src/tme/models/object_group_layer.rs
@@ -10,7 +10,7 @@ use super::property::Property;
 pub struct ObjectGroupLayer {
     #[serde(rename = "draworder")]
     pub draw_order: DrawOrder,
-    pub id:         i64,
+    pub id:         i32,
     pub name:       String,
     pub objects:    Vec<Object>,
     #[serde(rename = "offsetx")]
@@ -20,12 +20,12 @@ pub struct ObjectGroupLayer {
     pub opacity:    f64,
     pub properties: Option<Vec<Property>>,
     #[serde(rename = "startx")]
-    pub start_x:    Option<i64>,
+    pub start_x:    Option<i32>,
     #[serde(rename = "starty")]
-    pub start_y:    Option<i64>,
+    pub start_y:    Option<i32>,
     pub visible:    bool,
-    pub x:          i64,
-    pub y:          i64,
+    pub x:          i32,
+    pub y:          i32,
 }
 
 #[cfg(test)]
