+++ src/tme/models/tile.rs
@@ -9,17 +9,17 @@ use super::property::Property;
 #[serde(rename_all = "lowercase")]
 pub struct Tile {
     pub animation:    Option<Vec<Frame>>,
-    pub id:           i64,
+    pub id:           i32,
     pub image:        Option<String>,
     #[serde(rename = "imageheight")]
-    pub image_height: Option<i64>,
+    pub image_height: Option<i32>,
     #[serde(rename = "imagewidth")]
-    pub image_width:  Option<i64>,
+    pub image_width:  Option<i32>,
     #[serde(rename = "objectgroup")]
     pub object_group: Option<Layer>,
     pub probability:  Option<f64>,
     pub properties:   Option<Vec<Property>>,
-    pub terrain:      Option<Vec<i64>>,
+    pub terrain:      Option<Vec<i32>>,
     #[serde(rename = "type")]
     pub tile_type:    Option<String>,
 }
