+++ src/tme/models/tile_offset.rs
@@ -4,8 +4,8 @@ use serde::Serialize;
 #[derive(Debug, Serialize, Deserialize, Eq, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub struct TileOffset {
-    pub x: i64,
-    pub y: i64,
+    pub x: i32,
+    pub y: i32,
 }
 
 #[cfg(test)]
